chore(implementation): implement generic diff.
this include implementation with basic optimizations(less copying)
this is just a o(m*n) spaced naive implementation.
need to port all code in .cpp to .hpp or .h as this is a template
library.

TODO: make the implementation sutable for streaming, take input as 2
streams -> give changes as a stream.

this way we can make things easy for multithreadding (message passing).